# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DeepClean - Sui blockchain spam & phishing filter agent API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type ThreatVerdict = typeof ThreatVerdict[keyof typeof ThreatVerdict];


export const ThreatVerdict = {
  SAFE: 'SAFE',
  SUSPICIOUS: 'SUSPICIOUS',
  MALICIOUS: 'MALICIOUS',
} as const;

export type ThreatStatus = typeof ThreatStatus[keyof typeof ThreatStatus];


export const ThreatStatus = {
  quarantined: 'quarantined',
  released: 'released',
  burned: 'burned',
} as const;

export interface Threat {
  id: number;
  objectId: string;
  objectType: string;
  senderAddress: string;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  displayUrl?: string | null;
  riskScore: number;
  verdict: ThreatVerdict;
  reasonCode: number;
  confidence: number;
  flags: string[];
  reasoning: string;
  status: ThreatStatus;
  /** @nullable */
  walrusBlobId?: string | null;
  /** @nullable */
  quarantineTxDigest?: string | null;
  detectedAt: string;
  updatedAt?: string;
}

export interface ThreatAnalysisInput {
  objectId: string;
  objectType: string;
  senderAddress: string;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  displayUrl?: string | null;
  /** @nullable */
  moveAbi?: string | null;
}

export type ThreatAnalysisResultVerdict = typeof ThreatAnalysisResultVerdict[keyof typeof ThreatAnalysisResultVerdict];


export const ThreatAnalysisResultVerdict = {
  SAFE: 'SAFE',
  SUSPICIOUS: 'SUSPICIOUS',
  MALICIOUS: 'MALICIOUS',
} as const;

export interface ThreatAnalysisResult {
  riskScore: number;
  verdict: ThreatAnalysisResultVerdict;
  reasonCode: number;
  confidence: number;
  flags: string[];
  reasoning: string;
  /** @nullable */
  savedThreatId?: number | null;
}

export interface PopulateWalletBody {
  /** Sui wallet address to seed with synthetic spam objects */
  targetAddress: string;
  /** Optional real Sui TX digest if the caller already submitted a PTB */
  txDigest?: string | null;
}

export type PopulateWalletResultThreatsItem = {
  objectId: string;
  objectType: string;
  verdict: string;
  riskScore: number;
  /** @nullable */
  threatId?: number | null;
};

export interface PopulateWalletResult {
  /** Number of objects pushed through analysis */
  injected: number;
  /** Number of objects quarantined (risk_score >= 65) */
  quarantined: number;
  /**
     * Real Sui TX digest (null if synthetic-only run)
     * @nullable
     */
  txDigest: string | null;
  threats: PopulateWalletResultThreatsItem[];
}

export interface WatchedWallet {
  id: number;
  address: string;
  label: string;
  isActive: boolean;
  threatsDetected?: number;
  createdAt: string;
}

export interface WatchedWalletInput {
  address: string;
  label: string;
}

export interface DashboardStats {
  totalThreats: number;
  quarantined: number;
  burned: number;
  released: number;
  maliciousCount: number;
  suspiciousCount: number;
  safeCount: number;
  walletsMonitored: number;
  avgRiskScore: number;
}

export type ActivityEventType = typeof ActivityEventType[keyof typeof ActivityEventType];


export const ActivityEventType = {
  detected: 'detected',
  quarantined: 'quarantined',
  released: 'released',
  burned: 'burned',
} as const;

export interface ActivityEvent {
  id: number;
  type: ActivityEventType;
  objectId: string;
  objectType?: string;
  verdict: string;
  riskScore: number;
  timestamp: string;
}

export interface RiskBreakdownItem {
  reasonCode: number;
  label: string;
  count: number;
}

export type ListThreatsParams = {
verdict?: ListThreatsVerdict;
status?: ListThreatsStatus;
limit?: number;
offset?: number;
};

export type ListThreatsVerdict = typeof ListThreatsVerdict[keyof typeof ListThreatsVerdict];


export const ListThreatsVerdict = {
  SAFE: 'SAFE',
  SUSPICIOUS: 'SUSPICIOUS',
  MALICIOUS: 'MALICIOUS',
} as const;

export type ListThreatsStatus = typeof ListThreatsStatus[keyof typeof ListThreatsStatus];


export const ListThreatsStatus = {
  quarantined: 'quarantined',
  released: 'released',
  burned: 'burned',
} as const;

export type GetRecentActivityParams = {
limit?: number;
};