# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DeepClean - Sui blockchain spam & phishing filter agent API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from 'zod';


/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  "status": zod.string()
})


/**
 * @summary List all detected threats
 */
export const listThreatsQueryLimitDefault = 50;
export const listThreatsQueryOffsetDefault = 0;

export const ListThreatsQueryParams = zod.object({
  "verdict": zod.enum(['SAFE', 'SUSPICIOUS', 'MALICIOUS']).optional(),
  "status": zod.enum(['quarantined', 'released', 'burned']).optional(),
  "limit": zod.coerce.number().default(listThreatsQueryLimitDefault),
  "offset": zod.coerce.number().default(listThreatsQueryOffsetDefault)
})

export const ListThreatsResponseItem = zod.object({
  "id": zod.number(),
  "objectId": zod.string(),
  "objectType": zod.string(),
  "senderAddress": zod.string(),
  "displayName": zod.string().nullish(),
  "displayUrl": zod.string().nullish(),
  "riskScore": zod.number(),
  "verdict": zod.enum(['SAFE', 'SUSPICIOUS', 'MALICIOUS']),
  "reasonCode": zod.number(),
  "confidence": zod.number(),
  "flags": zod.array(zod.string()),
  "reasoning": zod.string(),
  "status": zod.enum(['quarantined', 'released', 'burned']),
  "walrusBlobId": zod.string().nullish(),
  "quarantineTxDigest": zod.string().nullish(),
  "detectedAt": zod.string(),
  "updatedAt": zod.string().optional()
})
export const ListThreatsResponse = zod.array(ListThreatsResponseItem)


/**
 * @summary Analyze a Sui object for threats using Gemini AI
 */
export const AnalyzeThreatBody = zod.object({
  "objectId": zod.string(),
  "objectType": zod.string(),
  "senderAddress": zod.string(),
  "displayName": zod.string().nullish(),
  "displayUrl": zod.string().nullish(),
  "moveAbi": zod.string().nullish()
})

export const AnalyzeThreatResponse = zod.object({
  "riskScore": zod.number(),
  "verdict": zod.enum(['SAFE', 'SUSPICIOUS', 'MALICIOUS']),
  "reasonCode": zod.number(),
  "confidence": zod.number(),
  "flags": zod.array(zod.string()),
  "reasoning": zod.string(),
  "savedThreatId": zod.number().nullish()
})


/**
 * @summary Get a specific threat by ID
 */
export const GetThreatParams = zod.object({
  "id": zod.coerce.number()
})

export const GetThreatResponse = zod.object({
  "id": zod.number(),
  "objectId": zod.string(),
  "objectType": zod.string(),
  "senderAddress": zod.string(),
  "displayName": zod.string().nullish(),
  "displayUrl": zod.string().nullish(),
  "riskScore": zod.number(),
  "verdict": zod.enum(['SAFE', 'SUSPICIOUS', 'MALICIOUS']),
  "reasonCode": zod.number(),
  "confidence": zod.number(),
  "flags": zod.array(zod.string()),
  "reasoning": zod.string(),
  "status": zod.enum(['quarantined', 'released', 'burned']),
  "walrusBlobId": zod.string().nullish(),
  "quarantineTxDigest": zod.string().nullish(),
  "detectedAt": zod.string(),
  "updatedAt": zod.string().optional()
})


/**
 * @summary Release a quarantined asset back to owner
 */
export const ReleaseThreatParams = zod.object({
  "id": zod.coerce.number()
})

export const ReleaseThreatResponse = zod.object({
  "id": zod.number(),
  "objectId": zod.string(),
  "objectType": zod.string(),
  "senderAddress": zod.string(),
  "displayName": zod.string().nullish(),
  "displayUrl": zod.string().nullish(),
  "riskScore": zod.number(),
  "verdict": zod.enum(['SAFE', 'SUSPICIOUS', 'MALICIOUS']),
  "reasonCode": zod.number(),
  "confidence": zod.number(),
  "flags": zod.array(zod.string()),
  "reasoning": zod.string(),
  "status": zod.enum(['quarantined', 'released', 'burned']),
  "walrusBlobId": zod.string().nullish(),
  "quarantineTxDigest": zod.string().nullish(),
  "detectedAt": zod.string(),
  "updatedAt": zod.string().optional()
})


/**
 * @summary Permanently burn a quarantined malicious asset
 */
export const BurnThreatParams = zod.object({
  "id": zod.coerce.number()
})

export const BurnThreatResponse = zod.object({
  "id": zod.number(),
  "objectId": zod.string(),
  "objectType": zod.string(),
  "senderAddress": zod.string(),
  "displayName": zod.string().nullish(),
  "displayUrl": zod.string().nullish(),
  "riskScore": zod.number(),
  "verdict": zod.enum(['SAFE', 'SUSPICIOUS', 'MALICIOUS']),
  "reasonCode": zod.number(),
  "confidence": zod.number(),
  "flags": zod.array(zod.string()),
  "reasoning": zod.string(),
  "status": zod.enum(['quarantined', 'released', 'burned']),
  "walrusBlobId": zod.string().nullish(),
  "quarantineTxDigest": zod.string().nullish(),
  "detectedAt": zod.string(),
  "updatedAt": zod.string().optional()
})


/**
 * @summary List all monitored wallets
 */
export const ListWatchedWalletsResponseItem = zod.object({
  "id": zod.number(),
  "address": zod.string(),
  "label": zod.string(),
  "isActive": zod.boolean(),
  "threatsDetected": zod.number().optional(),
  "createdAt": zod.string()
})
export const ListWatchedWalletsResponse = zod.array(ListWatchedWalletsResponseItem)


/**
 * @summary Add a wallet address to monitor
 */
export const AddWatchedWalletBody = zod.object({
  "address": zod.string(),
  "label": zod.string()
})


/**
 * @summary Remove a wallet from monitoring
 */
export const RemoveWatchedWalletParams = zod.object({
  "id": zod.coerce.number()
})

export const RemoveWatchedWalletResponse = zod.object({
  "id": zod.number(),
  "address": zod.string(),
  "label": zod.string(),
  "isActive": zod.boolean(),
  "threatsDetected": zod.number().optional(),
  "createdAt": zod.string()
})


/**
 * @summary Seed demo spam objects into a monitored wallet for testing
 */
export const PopulateWalletBody = zod.object({
  "targetAddress": zod.string().describe('Sui wallet address to seed with synthetic spam objects'),
  "txDigest": zod.string().nullish().describe('Optional real Sui TX digest if the caller already submitted a PTB')
})

export const PopulateWalletResponse = zod.object({
  "injected": zod.number().describe('Number of objects pushed through analysis'),
  "quarantined": zod.number().describe('Number of objects quarantined (risk_score >= 65)'),
  "txDigest": zod.string().nullable().describe('Real Sui TX digest (null if synthetic-only run)'),
  "threats": zod.array(zod.object({
  "objectId": zod.string(),
  "objectType": zod.string(),
  "verdict": zod.string(),
  "riskScore": zod.number(),
  "threatId": zod.number().nullish()
}))
})


/**
 * @summary Get dashboard summary statistics
 */
export const GetDashboardStatsResponse = zod.object({
  "totalThreats": zod.number(),
  "quarantined": zod.number(),
  "burned": zod.number(),
  "released": zod.number(),
  "maliciousCount": zod.number(),
  "suspiciousCount": zod.number(),
  "safeCount": zod.number(),
  "walletsMonitored": zod.number(),
  "avgRiskScore": zod.number()
})


/**
 * @summary Get recent threat activity feed
 */
export const getRecentActivityQueryLimitDefault = 10;

export const GetRecentActivityQueryParams = zod.object({
  "limit": zod.coerce.number().default(getRecentActivityQueryLimitDefault)
})

export const GetRecentActivityResponseItem = zod.object({
  "id": zod.number(),
  "type": zod.enum(['detected', 'quarantined', 'released', 'burned']),
  "objectId": zod.string(),
  "objectType": zod.string().optional(),
  "verdict": zod.string(),
  "riskScore": zod.number(),
  "timestamp": zod.string()
})
export const GetRecentActivityResponse = zod.array(GetRecentActivityResponseItem)


/**
 * @summary Get breakdown of threats by reason code
 */
export const GetRiskBreakdownResponseItem = zod.object({
  "reasonCode": zod.number(),
  "label": zod.string(),
  "count": zod.number()
})
export const GetRiskBreakdownResponse = zod.array(GetRiskBreakdownResponseItem)