# lib/api-zod/src/generated/types/activityEvent.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DeepClean - Sui blockchain spam & phishing filter agent API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityEventType } from './activityEventType';

export interface ActivityEvent {
  id: number;
  type: ActivityEventType;
  objectId: string;
  objectType?: string;
  verdict: string;
  riskScore: number;
  timestamp: string;
}