# lib/api-zod/src/generated/types/threat.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DeepClean - Sui blockchain spam & phishing filter agent API
 * OpenAPI spec version: 0.1.0
 */
import type { ThreatStatus } from './threatStatus';
import type { ThreatVerdict } from './threatVerdict';

export interface Threat {
  id: number;
  objectId: string;
  objectType: string;
  senderAddress: string;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  displayUrl?: string | null;
  riskScore: number;
  verdict: ThreatVerdict;
  reasonCode: number;
  confidence: number;
  flags: string[];
  reasoning: string;
  status: ThreatStatus;
  /** @nullable */
  walrusBlobId?: string | null;
  /** @nullable */
  quarantineTxDigest?: string | null;
  detectedAt: string;
  updatedAt?: string;
}